+++ packages/core/datetime-picker/src/components/DateTimePicker.js
@@ -32,6 +32,10 @@ type Props = {
   onFocus: () => void,
   /** The ISO time that should be used as the input value. */
   value?: string,
+  /** Allow users to edit the input and add a time */
+  timeIsEditable?: boolean,
+  /** Indicates current value is invalid & changes border color */
+  isInvalid?: boolean,
 };
 
 type State = {
@@ -43,14 +47,18 @@ type State = {
   zoneValue: string,
 };
 
+function getBorderColor(isInvalid: boolean) {
+  if (isInvalid) return `2px solid ${colors.R400}`;
+  return `1px solid ${colors.N20}`;
+}
 const Flex = styled.div`
   background-color: ${colors.N10};
   border-radius: ${borderRadius()}px;
   display: flex;
   transition: background-color 200ms ease-in-out, border-color 200ms ease-in-out;
-  ${({ isFocused }) => `
+  ${({ isFocused, isInvalid }) => `
     border: ${
-      isFocused ? `2px solid ${colors.B100}` : `1px solid ${colors.N20}`
+      isFocused ? `2px solid ${colors.B100}` : `${getBorderColor(isInvalid)}`
     };
     padding: ${isFocused ? '0' : '1px'};
   `} &:hover {
@@ -101,6 +109,8 @@ export default class DateTimePicker extends Component<Props, State> {
     innerProps: {},
     id: '',
     defaultValue: '',
+    timeIsEditable: false,
+    isInvalid: false,
   };
 
   state = {
@@ -167,15 +177,29 @@ export default class DateTimePicker extends Component<Props, State> {
   }
 
   render() {
-    const { autoFocus, id, innerProps, isDisabled, name } = this.props;
+    const {
+      autoFocus,
+      id,
+      innerProps,
+      isDisabled,
+      name,
+      timeIsEditable,
+    } = this.props;
     const { isFocused, value, dateValue, timeValue } = this.getState();
     const bothProps = {
       isDisabled,
       onBlur: this.onBlur,
       onFocus: this.onFocus,
+      isInvalid: this.props.isInvalid,
     };
+
     return (
-      <Flex {...innerProps} isFocused={isFocused} isDisabled={isDisabled}>
+      <Flex
+        {...innerProps}
+        isFocused={isFocused}
+        isDisabled={isDisabled}
+        isInvalid={bothProps.isInvalid}
+      >
         <input name={name} type="hidden" value={value} />
         <FlexItem>
           <DatePicker
@@ -194,7 +218,8 @@ export default class DateTimePicker extends Component<Props, State> {
             icon={CalendarIcon}
             onChange={this.onTimeChange}
             selectProps={{ styles }}
-            value={timeValue}
+            defaultValue={timeValue}
+            timeIsEditable={timeIsEditable}
           />
         </FlexItem>
       </Flex>