+++ packages/search/global-search/src/components/SearchResultsUtil.tsx
@@ -1,9 +1,11 @@
 import * as React from 'react';
 import { ComponentClass } from 'react';
-import { PersonResult, ResultBase } from '@atlaskit/quick-search';
-import ConfluenceIcon from '@atlaskit/icon/glyph/confluence';
+import {
+  PersonResult,
+  ContainerResult,
+  ResultBase,
+} from '@atlaskit/quick-search';
 import JiraIcon from '@atlaskit/icon/glyph/jira';
-import PeopleIcon from '@atlaskit/icon/glyph/people';
 import { Result, ResultType } from '../model/Result';
 import ObjectResult from './ObjectResult';
 
@@ -14,6 +16,9 @@ function getResultComponent(resultType: ResultType): ComponentClass {
     }
     case ResultType.Person: {
       return PersonResult;
+    }
+    case ResultType.Container: {
+      return ContainerResult;
     }
     default: {
       // Make the TS compiler verify that all enums have been matched
@@ -32,13 +37,21 @@ export function renderResults(results: Result[]) {
   });
 }
 
-export const searchConfluenceItem = (query: string) => (
+export interface AdvancedSearchItemProps {
+  query: string;
+  icon: JSX.Element;
+  text: string;
+}
+
+export const searchConfluenceItem = (props: AdvancedSearchItemProps) => (
   <ResultBase
-    href={`/wiki/dosearchsite.action?queryString=${encodeURIComponent(query)}`}
-    icon={<ConfluenceIcon size="large" label="Search Confluence" />}
+    href={`/wiki/dosearchsite.action?queryString=${encodeURIComponent(
+      props.query,
+    )}`}
+    icon={props.icon}
     key="search_confluence"
     resultId="search_confluence"
-    text="Search for more Confluence pages and blogs"
+    text={props.text}
   />
 );
 
@@ -52,13 +65,13 @@ export const searchJiraItem = (query: string) => (
   />
 );
 
-export const searchPeopleItem = () => (
+export const searchPeopleItem = (props: AdvancedSearchItemProps) => (
   <ResultBase
-    href="/home/people"
-    icon={<PeopleIcon size="large" label="Search People" />}
+    href={`/home/people?q=${encodeURIComponent(props.query)}`}
+    icon={props.icon}
     key="search_people"
     resultId="search_people"
-    text="Search for more people"
+    text={props.text}
   />
 );