+++ packages/search/global-search/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @atlaskit/global-search
 
+## 3.0.4
+- [patch] Link to people search includes query [6d2e946](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/6d2e946)
+
+## 3.0.3
+
+## 3.0.2
+- [patch] Rename EmptyState component to NoResults [cb73105](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/cb73105)
+
 ## 3.0.1
 - [patch] Simplify tests [d2437e9](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/d2437e9)